pyproject.toml
[tool.poetry]
name = "fiscal-de-musica"
version = "0.1.0"
description = "Tracker pessoal de músicas com integração Last.fm"
authors = ["Mariana Barreto <marianaportobarreto@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
psycopg = {extras = ["binary"], version = "^3.2.0"}
httpx = "^0.28.0"
python-dotenv = "^1.0.0"
pydantic = "^2.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
